Move `encode_image()` to common for MIT licensing
Re-exported from previous location
