+++ src/libertem/viz/base.py
@@ -1,13 +1,13 @@
 import logging
-from io import BytesIO
 import time
-from typing import Dict, Optional
 
 import numpy as np
 from matplotlib import colors, cm
-from PIL import Image
 
 from libertem.udf.base import UDFRunner
+
+# Import here for backwards compatibility, refs #1031
+from libertem.common.viz import encode_image  # NOQA: F401
 
 logger = logging.getLogger(__name__)
 
@@ -32,37 +32,6 @@ def _get_norm(result, norm_cls=colors.Normalize, vmin=None, vmax=None, damage=No
         vmax = np.max(result[damage])
 
     return norm_cls(vmin=vmin, vmax=vmax)
-
-
-def encode_image(result, save_kwargs: Optional[Dict] = None) -> BytesIO:
-    """
-    Save the RGBA data in ``result`` to an image with parameters ``save_kwargs``
-    passed to ``PIL.Image.save``.
-
-    Parameters
-    ----------
-    result : numpy.ndarray
-        Array of RGB values; shape (height, width, 3)
-
-    save_kwargs : dict or None
-        dict of kwargs passed to Pillow when saving the image, can be used to set
-        the file format, quality, ...
-
-    Returns
-    -------
-
-    BytesIO
-        a buffer containing the result image (as PNG/JPG/... depending on save_kwargs)
-    """
-    if save_kwargs is None:
-        save_kwargs = {'format': 'png'}
-    # see also: https://stackoverflow.com/a/10967471/540644
-    im = Image.fromarray(result)
-    buf = BytesIO()
-    im = im.convert(mode="RGB")
-    im.save(buf, **save_kwargs)
-    buf.seek(0)
-    return buf
 
 
 def visualize_simple(result, colormap=None, logarithmic=False, vmin=None, vmax=None, damage=None):
